Reformatting of semantics
The goal is to have a better view of the definition of each FFV1 bitstream field, including the definition of all acceptable values per field and a list of the reserved values for future use.
It is as much as possible a copy of the content already in the document.