Reduce redundancy in the description of xxPlane() and xxLine().
LumaPlane(), CbPlane(), CrPlane() and AlphaPlane() are actually same;
the order of planes is already defined in the General section and has no impact
on the bitstream. Reduced to one Plane( p ) call.
LumaLine(), CbLine(), CrLine() and AlphaLine() are actually same;
the order of lines is already defined in the General section and has no impact
on the bitstream. Reduced to one Line( p, y ) call.
plane_count name may be misleading (it is the count of quant_table_index,
which is not always the count of planes) and does not exist in the bistream,
replaced by the sum of existing bitstream elements.
colorspace_type related "if" sorted in ascending order.