Correct extension in ConfigurationRecord
This issue was introduced very long ago (ebe963b70474e2f850223d93cce97d82c3b95990)
The range coder must know the size in this specific case to do the un-termination
Thus it is not possible to add symbols after the termination but it should be
possible to add them before. So that is the location where extensions should go

I am not 100% satisfied with the way this is specified in this change, so if someone
sees a better way to specify this, contact me, send a pull req.

Signed-off-by: Michael Niedermayer <michael@niedermayer.cc>