ffv1: Support CRC with initial and final values not 0
This allows detection of appended and prepended 0 bits.
Previously a correct block could have 0 bits added before and
after it and this would not be detected with the CRC.

I am not aware of such a case actually occurring in the real world but this
should improve the performance of the checksum slightly

The use of a random constant as initial and final values instead
of -1 (or 0) should improve the robustness in 2 ways:
1. a message starting with 32 1 bits would bring the CRC state to 0 and thus
again having undetectable 0 bit insertions.
while this is a constructed case and one can always construct such a case, using a
random initial value should make the pattern triggering this less simple.

2. -1 is commonly used like also 0, so if two data streams are mixed, it is conceivable that non ffv1
data could end in a ffv1 stream and not be detected as crc failures, this is again very constructed

Using a random constant instead of -1 seems to have no real cost, so even with negligible
gains it is better to pick a random constant.

Signed-off-by: Michael Niedermayer <michael@niedermayer.cc>